pyproject.toml
[project]
name = "syn-via-diff-2-real"
version = "0.1.0"
description = "Synthetic-to-real diffusion bridge pipeline and tools."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clean-fid>=0.1.35",
    "diffusers>=0.24.0",
    "numpy>=1.24",
    "pillow>=10.0",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "rich>=13.0",
    "safetensors>=0.4.0",
    "torch>=2.1.0",
    "typer>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]
